successfully retrieve and cache images
However cells information is getting misplaced and sometimes images are not loading.

images not loading could be related to how i go about cancelling the request and how to use it with prepareForReuse, but no clear solution is present. Issue persists whether using my own method alamofire imageView extension
whats also strange however, is text is getting mismatched. I'm not sure if this is because i query for the user name also or if my datasource of plans array and users dictionary is somehow getting mismatched. Need further review